+++ test/mangle.ts
@@ -1,1 +1,2 @@
+// node dist/bin.mjs --mangle-props=_$ test/mangle.ts
 console.log({ foo_: 0 });