+++ docs/.vitepress/config.ts
@@ -5,9 +5,10 @@ export default defineConfig({
   description: "A simple wrapper of esbuild to run your script file.",
   base: "/esbuild-dev/",
   themeConfig: {
-    // repo: "hyrious/esbuild-dev",
-    // docsDir: "docs",
-    // editLinks: true,
+    socialLinks: [
+      { icon: "github", link: "https://github.com/hyrious/esbuild-dev" },
+    ],
+
     editLink: {
       pattern: "https://github.com/hyrious/esbuild-dev/edit/main/docs/:path",
     },